Moving kepler into feature folder
