+++ examples/worldview/src/constants.js
@@ -1,0 +1,4 @@
+export const AUTH_TOKENS = {
+  MAPBOX_TOKEN: process.env.MapboxAccessToken, // eslint-disable-line
+  EXPORT_MAPBOX_TOKEN: process.env.MapboxExportToken // eslint-disable-line
+};
