+++ examples/worldview/src/features/kepler/constants.js
@@ -1,0 +1,1 @@
+export const LOADING_SAMPLE_ERROR_MESSAGE = 'Not able to load sample';
